refactor: split profile group block
